refactor: separate search functionality into cmd/search.go from main.go
Follow Cobra's recommended folder structure example.
